ai-control/requirements.txt
fastapi
python-socketio
httpx
anthropic
google-generativeai
boto3
python-dotenv
uvicorn
aiohttp
python-json-logger
typing-extensions